requirements.txt
streamlit
python-docx
langchain
langchain-openai
faiss-cpu
tiktoken
pandas
python-dotenv
openai
dotenv